course page fixed
