--- src/app/instructor/courses/page.tsx
@@ -1,141 +1,0 @@
-'use client';
-import React, { useState } from 'react';
-
-interface Course {
-  id: string;
-  name: string;
-  code: string;
-  description: string;
-  startDate: string;
-  endDate: string;
-  category: string;
-  students: { id: string; name: string; email: string }[];
-}
-
-const initialCourses: Course[] = [
-  {
-    id: 'ENG',
-    name: 'Diploma in English',
-    code: 'OXF/ENG/01',
-    description: 'English language basics and grammar.',
-    startDate: '2024-06-01',
-    endDate: '2024-12-01',
-    category: 'Language',
-    students: [
-      { id: 'S1', name: 'Alex John', email: 'alex@example.com' },
-      { id: 'S2', name: 'Sara Kim', email: 'sara@example.com' },
-    ],
-  },
-];
-
-export default function CourseManagementPage() {
-  const [courses, setCourses] = useState<Course[]>(initialCourses);
-  const [form, setForm] = useState<Omit<Course, 'students'>>({
-    id: '',
-    name: '',
-    code: '',
-    description: '',
-    startDate: '',
-    endDate: '',
-    category: '',
-  });
-  const [isEditing, setIsEditing] = useState(false);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    setForm({ ...form, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = () => {
-    if (!form.id || !form.name || !form.code) return;
-    if (isEditing) {
-      setCourses(courses.map((c) =>
-        c.id === form.id ? { ...form, students: c.students } : c
-      ));
-    } else {
-      setCourses([...courses, { ...form, students: [] }]);
-    }
-
-    resetForm();
-  };
-
-  const handleEdit = (course: Course) => {
-    const { students, ...courseInfo } = course;
-    setForm(courseInfo);
-    setIsEditing(true);
-  };
-
-  const handleDelete = (id: string) => {
-    setCourses(courses.filter((c) => c.id !== id));
-  };
-
-  const resetForm = () => {
-    setForm({ id: '', name: '', code: '', description: '', startDate: '', endDate: '', category: '' });
-    setIsEditing(false);
-  };
-
-  return (
-    <div className="bg-[#ecf3f4] min-h-screen p-6">
-      <div className="max-w-5xl mx-auto">
-        <h1 className="text-2xl font-bold text-[#194e63] mb-6">Course Management</h1>
-
-        <div className="bg-white p-6 rounded-lg shadow mb-10">
-          <h2 className="text-lg font-semibold mb-4 text-gray-700">
-            {isEditing ? 'Edit Course' : 'Add New Course'}
-          </h2>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <input name="id" placeholder="Course ID" value={form.id} onChange={handleChange} className="border rounded p-2" disabled={isEditing} />
-            <input name="name" placeholder="Course Name" value={form.name} onChange={handleChange} className="border rounded p-2" />
-            <input name="code" placeholder="Course Code" value={form.code} onChange={handleChange} className="border rounded p-2" />
-            <input name="category" placeholder="Category" value={form.category} onChange={handleChange} className="border rounded p-2" />
-            <input name="startDate" type="date" placeholder="Start Date" value={form.startDate} onChange={handleChange} className="border rounded p-2" />
-            <input name="endDate" type="date" placeholder="End Date" value={form.endDate} onChange={handleChange} className="border rounded p-2" />
-            <textarea name="description" placeholder="Course Description" value={form.description} onChange={handleChange} className="border rounded p-2 col-span-full" />
-          </div>
-          <div className="mt-4 flex gap-2">
-            <button onClick={handleSubmit} className="bg-[#194e63] text-white px-4 py-2 rounded shadow">
-              {isEditing ? 'Update Course' : 'Add Course'}
-            </button>
-            {isEditing && (
-              <button onClick={resetForm} className="px-4 py-2 rounded border border-gray-300 bg-gray-100 text-gray-600">
-                Cancel
-              </button>
-            )}
-          </div>
-        </div>
-
-        <div className="bg-white p-6 rounded-lg shadow">
-          <h2 className="text-lg font-semibold mb-4 text-gray-700">Courses</h2>
-          {courses.length > 0 ? (
-            <table className="w-full table-auto">
-              <thead className="text-left text-gray-600 border-b">
-                <tr>
-                  <th className="p-3">ID</th>
-                  <th className="p-3">Name</th>
-                  <th className="p-3">Code</th>
-                  <th className="p-3">Enrolled</th>
-                  <th className="p-3">Actions</th>
-                </tr>
-              </thead>
-              <tbody>
-                {courses.map((course) => (
-                  <tr key={course.id} className="border-b hover:bg-gray-50">
-                    <td className="p-3">{course.id}</td>
-                    <td className="p-3">{course.name}</td>
-                    <td className="p-3">{course.code}</td>
-                    <td className="p-3">{course.students.length}</td>
-                    <td className="p-3 flex gap-2">
-                      <button onClick={() => handleEdit(course)} className="text-sm text-blue-600">Edit</button>
-                      <button onClick={() => handleDelete(course.id)} className="text-sm text-red-600">Delete</button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          ) : (
-            <p className="text-gray-500">No courses available.</p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
